==> 02-rearrange-data.py <==
import json,glob
from tools import start_debugger_on_exception
from pathlib import Path
import re
from text_utils import chinese_ratio

start_debugger_on_exception()
def split_para(x):
    if '|' in x:
        return x.split('|')
    elif '\n' in x:
        import pdb;pdb.set_trace()
        return x.split('\n')
    #elif '    ' in x:
    #    return x.split('    ')
    #elif '\\' in x and ('"'  in x or '“' not in x and '”' not in x:
    #    import pdb;pdb.set_trace()
    else:
        return [x]
def len_ch_check(x):
    if x != '':
        if  len(x)<=10:
            x = ''
        elif chinese_ratio(x)<0.6:
            x = ''
    return x
def clean_data(x):
    head_no = u'?？。，,”")）》'
    tail_no = u'"“《（('
    ilegal_char = u'[^\u4E00-\u9FA5?？!.！、\;；<>\[\]【】\{\}!@#$%^&*()！@#¥%&（）～·`~、\-_——\-+=:：……0-9a-zA-Z。，,”")）》"“《（(|\/「」\'’‘．\n]'
    r_second_modify = re.compile(u'[^'+head_no+u'].*[^'+tail_no+u']',re.UNICODE)
    start = u'[^\u4E00-\u9FA5]{,30}'
    end = u'[^\u4E00-\u9FA5]{,30}'
    list_rep = [re.compile(start + u'小[^\u4E00-\u9FA5]{,20}说[^\u4E00-\u9FA5]{,20}天[^\u4E00-\u9FA5]{,20}堂'+end, re.UNICODE),\
                re.compile(start+u'[wWＷ][^\u4E00-\u9FA5]{,20}[wWＷ][^\u4E00-\u9FA5]{,20}[wWＷ][^\u4E00-\u9FA5]{,30}$', re.UNICODE),
                re.compile(start +u'[ＣcC][^\u4E00-\u9FA5]{,20}[oO][^\u4E00-\u9FA5]{,20}[mMＭ]'+end, re.UNICODE),
                re.compile(start +u'[tT][^\u4E00-\u9FA5]{,20}[xX][^\u4E00-\u9FA5]{,20}[tT][^\u4E00-\u9FA5]{,20}天[^\u4E00-\u9FA5]{,20}堂'+end, re.UNICODE),
                re.compile(start+u'xiaoshuotxt'+end,re.UNICODE),
                re.compile(u'mpanel(1);',re.UNICODE)
                ]
    list_del = [re.compile(u'读后感', re.UNICODE )]
    for rs in list_rep:
        ms = re.findall(rs,x)
        for m in ms:
            mm = re.search(r_second_modify,m).group()
           
            x = x.replace(mm,'')
    for rs in list_del:
        if re.search(rs,x) is not None:
            x = ''
    wrong_delimit = re.findall(re.compile(u'[^?!.？。！……"”\.\!\?]\|',re.UNICODE),x)
    for m in wrong_delimit:
        x = x.replace(m,m[:-1])
    if x != '':
        if  len(x)<=10:
            x = ''
        elif chinese_ratio(x)<0.6:
            x = ''
    if "\\" in x and '"' not in x and '“' not in x and '”' not in x:
        #import pdb;pdb.set_trace()
        x= x.replace('\\','"')
    x = x.replace('「','“')
    x= x.replace('」','”')
    #x = x.replace('“','"').replace('”','"')
    x = x.replace(' ','')
    # x = x.replace('\t','')
    # if len(re.findall(ilegal_char,x))>0:

    #     import pdb;pdb.set_trace()
    x = re.sub(ilegal_char,'',x,flags=re.U)
    return x
authors = [f.split('/')[-1].split('.')[0] for f in glob.glob('00-grabdata/*.json')]
def write_data_title_paragraph(author):
    book_per_author = json.load(open('01-name-the-books-per-author.json','r'))
    f = '00-grabdata/'+author+'.json'
    data = json.load(open(f,'r'))
    output = {}
    leng = 0
    paras = 0
    docs = 0
    chaps = 0
    for title in data.keys():
        if title+'DEL' in book_per_author[author]:
            continue
        docs += 1
        output[title] = []
        chapter = data[title]
        for c in chapter:
            if 'content' in c.keys():
                content_str = 'content'
            elif 'text' in c.keys():
                content_str = 'text'
            c[content_str] = clean_data(c[content_str]) if isinstance(c[content_str],str) else [clean_data(x) for x in c[content_str]]
            if  isinstance(c[content_str],str):
                c['content'] = split_para(c[content_str])
            elif isinstance(c[content_str],list):
                c['content'] = c[content_str]
            else:
                assert False
            chaps +=1
            paras += len(c['content'])
            leng += sum([len(x) for x in c['content']])
            temp = [x.strip() for x in c['content'] if len(x.strip())>1  and len_ch_check(x.strip()) != '']
            #temp = sum([cut_string_zh(x,300) for x in temp],[])

            output[title]+=temp
    print('author:',author)
    print('    ','docs:',docs)
    print('    ','chaps:',chaps)
    print('    ','length:',leng)
    path = '00-grabdata/'+author
    Path(path).mkdir(parents=True, exist_ok=True)
    with open(path+'/data.json','w') as f:
        json.dump(output,f,ensure_ascii=False,indent=4)
for author in authors:    
    write_data_title_paragraph(author=author)
